Move tests to test_dependency_ordering.py
